% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{db_path}
\alias{db_path}
\title{Fech visualization data}
\format{
An object of class \code{character} of length 1.
}
\usage{
db_path
}
\description{
Fech visualization data
}
\keyword{datasets}
