% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{display_selectize}
\alias{display_selectize}
\title{Function to display selectize input}
\usage{
display_selectize(id, items, selectize.title = NULL)
}
\arguments{
\item{id}{is the id of each selectize}

\item{items}{vector of options. The first item will be used as default.}

\item{selectize.title}{string to provide a title}
}
\description{
Function to display selectize input
}
