% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visualization_utils_.R
\name{icc_viewer}
\alias{icc_viewer}
\title{Function to display ICC and IF images from Human Protein Atlas}
\usage{
icc_viewer(icc_data)
}
\arguments{
\item{title}{is a vector of string indicating the titles}

\item{named}{list indicating the content of each tabpanel. The list names correspond to the vector of title.}

\item{id}{is the id of the tab}
}
\description{
Function to display ICC and IF images from Human Protein Atlas
}
