% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadDataTrans}
\alias{loadDataTrans}
\title{Load expression data of specific transcript}
\usage{
loadDataTrans(ens_l)
}
\arguments{
\item{ens_l}{vector of Ensembl transcript id}
}
\description{
Load expression data of specific transcript
}
