% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visualization_utils_.R
\name{tabpanel2render}
\alias{tabpanel2render}
\title{Function to display tabpanel}
\usage{
tabpanel2render(id, title, content, gene.symbol)
}
\arguments{
\item{id}{is the id of the tab}

\item{title}{is a vector of string indicating the titles}

\item{gene.symbol}{selected gene symbol}

\item{named}{list indicating the content of each tabpanel. The list names correspond to the vector of title.}
}
\description{
Function to display tabpanel
}
