% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visualization_utils_.R
\name{viz.helper}
\alias{viz.helper}
\title{Function to reuse the rendering of visualization page}
\usage{
viz.helper(sel_gene, top_ref = 10, specie, ns, output, input)
}
\arguments{
\item{sel_gene}{the selected gene or gene name retrieved from url parameter "gene"}

\item{top_ref}{is the number of top rank of reference transcript. Check the function get_gene_info}

\item{specie}{string indicating the specie}

\item{ns}{function. name space of the module}
}
\description{
Function to reuse the rendering of visualization page
}
