% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visualization_utils_.R
\name{hk_transcript_desc}
\alias{hk_transcript_desc}
\title{Function to define render the description of each transcript}
\usage{
hk_transcript_desc(
  specie,
  gene.symbol,
  gene.name,
  gene.description,
  transcript,
  tissue_info,
  gene.other.names = NULL,
  ortholog = NULL
)
}
\arguments{
\item{specie}{human or mouse}

\item{gene.symbol}{the symbol of selected gene. Default is the default name in selectize}

\item{gene.name}{is the official name of the seclected gene}

\item{transcript}{string or vector of string of transcripts. If length is > 1,
all of the transcripts will be displayed if the target tissues are different.}

\item{tissue_info}{tissues for which the transcript has been proposed
to be a reference transcript
ortholog the second specie if the same gene has been idenfity as HK in both}

\item{gene.other.names}{Other known names of the gene. It will appear if not Null}
}
\description{
Function to define render the description of each transcript
}
